server/requirements.txt
asyncio
websockets
python-dotenv
google-generativeai
nltk
